fix: wrong duration on rebook modals
