+++ src/slices/hatch/booking-page/components/modals/WeekModal.tsx
@@ -7,6 +7,7 @@ import {
   useAddRoomBookingHook,
   useFetchAvailabilitiesHook,
 } from '@/slices/hatch/booking-page/hooks/bookingHooks';
+import { getDuration } from '@/slices/hatch/booking-page/utils';
 
 type WeekModalProps = {
   open: boolean;
@@ -104,17 +105,6 @@ const WeekModal: React.FC<WeekModalProps> = ({
       },
     });
     onClose();
-  }
-
-  function getDuration(startTime: Date, endTime: Date) {
-    const timeMin = differenceInMinutes(endTime, startTime);
-    const hours = Math.floor(timeMin / 60);
-    const min = timeMin % 60;
-    if (min != 0) {
-      return hours + 0.5;
-    } else {
-      return hours;
-    }
   }
 
   function isAvail(
